backend/requirements.txt
Flask==3.1.2
icecream==2.1.10
mysql-connector-python==9.0.0
Werkzeug==3.1.6
PyJWT==2.9.0
flask-cors==4.0.0